Fix Numpy <---> TensorProto Conversion (#60)
* Fixed conversion for scalar types, both from and to TensorProto
* Fixed conversion from nested python lists
* Fixed conversion for DT_STRING, DT_HALF, DT_COMPLEX64 and DT_COMPLEX128
* Added tests to test Numpy to TensorProto conversion
* Fix tests
* Bumped version from 2.3.1 to 2.3.2

+++ tests/test_predictor.py
@@ -145,7 +145,7 @@ def test_predict_python_scalar_type(scalar_servable):
     predictions = predictor_client.predict(inputs=inputs)
 
     assert isinstance(predictions, dict)
-    assert isinstance(predictions['output'], int)
+    assert isinstance(predictions['output'], np.int64)
     assert predictions['output'] == value
